feat: created a new file called hooks and useAuth.ts to verify the token
